fix: resolve crucially errors in ActivationCodeScreen, DependentInfoScreen, SettingsScreen, PersonalProfileScreen, ParentDeletePermission and related components.

+++ src/Screens/Main/ParentDeletePermission/ParentDeletePermissionScreen.tsx
@@ -1,29 +1,30 @@
-import React, { useState } from "react";
-import { useNavigation, useRoute } from "@react-navigation/native";
-import { StyleSheet, View } from "react-native";
-import { Text, Button, Select, SelectItem } from "@ui-kitten/components";
-import { AppHeader } from "@/Components";
-import Colors from "@/Theme/Colors";
-import { loadUserId } from "@/Storage/MainAppStorage";
-import { DeleteStudent } from "@/Services/Student";
+import React, { useState } from 'react';
+import { RouteProp, useNavigation, useRoute } from '@react-navigation/native';
+import { StyleSheet, View } from 'react-native';
+import { Text, Button, Select, SelectItem } from '@ui-kitten/components';
+import { AppHeader } from '@/Components';
+import Colors from '@/Theme/Colors';
+import { loadUserId } from '@/Storage/MainAppStorage';
+import { DeleteStudent } from '@/Services/Student';
+import { MainStackNavigatorParamsList } from '@/Navigators/Main/RightDrawerNavigator';
 
-const permissions = ["Yes", "No"];
+const permissions = ['Yes', 'No'];
 
 const ParentDeletePermissionScreen = () => {
   const navigation = useNavigation();
-  const route = useRoute();
+  const route = useRoute<RouteProp<MainStackNavigatorParamsList, 'ParentDeletePermission'>>();
   const [selectedItem, setSelectedItem] = useState(1);
 
   return (
     <>
       <AppHeader title="Delete Permission" isBack />
       <View style={{ padding: 10, marginTop: 20 }}>
-        <Text style={{ fontSize: 16, fontWeight: "600", width: "100%" }}>
+        <Text style={{ fontSize: 16, fontWeight: '600', width: '100%' }}>
           Are you sure you want to give permission to delete account?
         </Text>
       </View>
       <Select
-        style={{ width: "90%", marginHorizontal: "5%" }}
+        style={{ width: '90%', marginHorizontal: '5%' }}
         value={permissions[selectedItem]}
         placeholder="Grade"
         onSelect={(index: any) => {
@@ -40,9 +41,10 @@ const ParentDeletePermissionScreen = () => {
         size="small"
         onPress={async () => {
           const userId = await loadUserId();
+          if (!userId) return;
           DeleteStudent(route?.params?.dependentId, userId)
             .then((response) => {
-              console.log("response", response);
+              console.log('response', response);
               navigation.goBack();
             })
             .catch((error) => console.log(error));
@@ -59,25 +61,25 @@ export default ParentDeletePermissionScreen;
 const styles = StyleSheet.create({
   layout: {
     flex: 1,
-    flexDirection: "column",
+    flexDirection: 'column',
     paddingTop: 50,
   },
   modalButton: {
-    width: "90%",
+    width: '90%',
     marginTop: 30,
-    marginHorizontal: "5%",
+    marginHorizontal: '5%',
   },
   item: {
     borderRadius: 10,
-    width: "96%",
-    backgroundColor: "#fff",
+    width: '96%',
+    backgroundColor: '#fff',
     marginVertical: 10,
-    marginHorizontal: "2%",
+    marginHorizontal: '2%',
     padding: 10,
   },
   row: {
-    flexDirection: "row",
-    alignItems: "center",
+    flexDirection: 'row',
+    alignItems: 'center',
   },
   text: {
     fontSize: 16,
@@ -89,15 +91,15 @@ const styles = StyleSheet.create({
     borderColor: Colors.primary,
     height: 30,
     width: 40,
-    alignItems: "center",
-    justifyContent: "center",
+    alignItems: 'center',
+    justifyContent: 'center',
   },
   textButton: {
     color: Colors.primary,
     fontSize: 16,
   },
   floatButton: {
-    position: "absolute",
+    position: 'absolute',
     bottom: 20,
     right: 20,
     shadowColor: Colors.primary,
